django/requirements.txt
asgiref==3.2.3
certifi==2020.12.5
chardet==4.0.0
Django==3.0.3
django-cors-headers==3.7.0
django-hCaptcha==0.1.0
djangorestframework==3.12.4
gunicorn==20.0.4
idna==2.10
mysqlclient==2.0.3
pytz==2019.3
requests==2.25.1
sqlparse==0.3.0
urllib3==1.26.4
